added copyright and spec implementation
