+++ myosuite/envs/myo/assets/arm/myoarm_tabletennis.xml
@@ -1,4 +1,12 @@
 <mujoco model="ping_pong_paddle">
+<!-- =================================================
+    Copyright (c) MyoSuite Authors
+    Model   :: Myo PingPong (MuJoCoV3.3)
+    Author  :: Cheryl Wang (huiyi.wang@mail.mcgill.ca), Balint Hodossy(bkh16@ic.ac.uk), Vikash Kumar (vikashplus@gmail.com), Vittorio Caggiano
+    source  :: https://github.com/MyoHub/MyoSuite
+    License :: Under Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+====================================================== -©vk©-->
+
     <include file="../../../../simhive/myo_sim/torso/assets/myotorso_assets.xml"/>
     <include file="../../../../simhive/myo_sim/head/assets/myohead_simple_assets.xml"/>
     <include file="../../../../simhive/myo_sim/leg/assets/myolegs_assets.xml"/>
